EXWM: Disable the dynamic terminal name
It seems an EXWM buffer renaming itself while on another workspace
causes it to show up on the other workspace too
temporarily (apparently until the original workspace is focused). The
terminal window is especially prone to this, especially while running
an emerge(1) process which displays the current system load. Due to
this, I decided to disable this feature.