Add a dedicated org-capture inbox file
Adding new entries effectively in the middle of gtd.org often caused
me to accidentally remove the newline separating the new entry from
the following top-level heading, wreaking havoc until noticed.

It should no longer be a problem now that this new file is strictly an
inbox so I always append to it.