org-agenda: Don't show the clocked time after pressing "l"
It can be forced by the prefix argument, which by default doesn't
change much as the clocked time generates the most noise.  After this
change the default view only includes the task close times (which are
useful for quick reports of what was done) and more info can be
requested with the prefix argument, as stated before.