Prevent skeletons from being loaded multiple times
Multiple calls to define-auto-insert, even with the same argument, was
causing the skeleton to be inserted multiple times.

Previously require was replaced with load to ensure the whole
initialization is "reproducible" during runtime, but in this case it
shouldn't hurt as skeletons are quite simple (when compared to
installing and settings up multiple 3rd party packages).